# Improve find performance